pwm fix

+++ custom_board/arduino/cores/riscduino/wiring_analog.c
@@ -109,7 +109,7 @@ void analogWrite(uint32_t pin, uint32_t ulValue)
        pwm_enabled[pwm_num] = 0;
     } else {
        //*((volatile uint16_t*) (pwm_num*4 + PWM_BASE_ADDR +PWM_CFG_HIGH_BASE))   = ulValue & 0xFF;
-       *((volatile uint16_t*) (pwm_num*4 + PWM_BASE_ADDR +PWM_CFG_LOW_BASE))   = ((ulValue & 0xFF) << 16) | ((255-ulValue) & 0xFF);
+       *((volatile uint32_t*) (pwm_num*4 + PWM_BASE_ADDR +PWM_CFG_LOW_BASE))   = ((ulValue & 0xFF) << 16) | ((255-ulValue) & 0xFF);
     }
